feat: ca available in Widgets (DAP-4841)
